# docs/swagger.yaml
definitions:
  entity.Accounts:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
  entity.Categories:
    properties:
      categoryId:
        type: integer
      categoryName:
        type: string
      description:
        type: string
      picture:
        items:
          type: integer
        type: array
    type: object
info:
  contact: {}
paths:
  /v1/account/create:
    post:
      consumes:
      - application/json
      description: save Account
      parameters:
      - description: Account
        in: body
        name: body
        schema:
          $ref: '#/definitions/entity.Accounts'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.Accounts'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Account
      tags:
      - Account
  /v1/categories/{id}:
    get:
      consumes:
      - application/json
      description: get category
      parameters:
      - description: category_id
        in: path
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.Categories'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Category
      tags:
      - Category
  /v1/categories/create:
    post:
      consumes:
      - application/json
      description: save category
      parameters:
      - description: category
        in: body
        name: body
        schema:
          $ref: '#/definitions/entity.Categories'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.Categories'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Category
      tags:
      - Category
  /v1/categories/delete/{id}:
    delete:
      consumes:
      - application/json
      description: delete category
      parameters:
      - description: id
        in: path
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.Categories'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Category
      tags:
      - Category
  /v1/categories/list:
    get:
      consumes:
      - application/json
      description: get list category
      parameters:
      - description: offset
        in: header
        name: offset
        type: integer
      - description: limit
        in: header
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.Categories'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: ListCategory
      tags:
      - Category
  /v1/categories/update:
    patch:
      consumes:
      - application/json
      description: update category
      parameters:
      - description: category
        in: body
        name: body
        schema:
          $ref: '#/definitions/entity.Categories'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.Categories'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Category
      tags:
      - Category
swagger: "2.0"